tests: allow static compilation
