+++ Source/Lib/Decoder/Codec/EbDecBitstreamUnit.h
@@ -126,14 +126,166 @@ typedef struct OdEcDec {
     int16_t cnt;
 } OdEcDec;
 
-int od_ec_decode_bool_q15(OdEcDec *dec, unsigned f);
-int od_ec_decode_cdf_q15(OdEcDec *dec, const uint16_t *cdf, int nsyms);
+static void od_ec_dec_refill(OdEcDec *dec);
+
+/*Takes updated dif and range values, renormalizes them so that
+   32768 <= rng < 65536 (reading more bytes from the stream into dif if
+   necessary), and stores them back in the decoder context.
+  dif: The new value of dif.
+  rng: The new value of the range.
+  ret: The value to return.
+  Return: ret.
+          This allows the compiler to jump to this function via a tail-call.*/
+static int od_ec_dec_normalize(OdEcDec *dec, OdEcWindow dif, unsigned rng, int ret) {
+    int d;
+    assert(rng <= 65535U);
+    /*The number of leading zeros in the 16-bit binary representation of rng.*/
+    d = 16 - OD_ILOG_NZ(rng);
+    /*d bits in dec->dif are consumed.*/
+    dec->cnt -= d;
+    /*This is equivalent to shifting in 1's instead of 0's.*/
+    dec->dif = ((dif + 1) << d) - 1;
+    dec->rng = rng << d;
+    if (dec->cnt < 0)
+        od_ec_dec_refill(dec);
+    return ret;
+}
+
+/*Decode a single binary value.
+  f: The probability that the bit is one, scaled by 32768.
+  Return: The value decoded (0 or 1).*/
+static int od_ec_decode_bool_q15(OdEcDec *dec, unsigned f) {
+    OdEcWindow dif;
+    OdEcWindow vw;
+    unsigned   r;
+    unsigned   r_new;
+    unsigned   v;
+    int        ret;
+    assert(0 < f);
+    assert(f < 32768U);
+    dif = dec->dif;
+    r   = dec->rng;
+    assert(dif >> (OD_EC_WINDOW_SIZE - 16) < r);
+    assert(32768U <= r);
+    v = ((r >> 8) * (uint32_t)(f >> EC_PROB_SHIFT) >> (7 - EC_PROB_SHIFT));
+    v += EC_MIN_PROB;
+    vw    = (OdEcWindow)v << (OD_EC_WINDOW_SIZE - 16);
+    ret   = 1;
+    r_new = v;
+    if (dif >= vw) {
+        r_new = r - v;
+        dif -= vw;
+        ret = 0;
+    }
+    return od_ec_dec_normalize(dec, dif, r_new, ret);
+}
+
+/*Decodes a symbol given an inverse cumulative distribution function (CDF)
+   table in Q15.
+  icdf: CDF_PROB_TOP minus the CDF, such that symbol s falls in the range
+         [s > 0 ? (CDF_PROB_TOP - icdf[s - 1]) : 0, CDF_PROB_TOP - icdf[s]).
+        The values must be monotonically non-increasing, and icdf[nsyms - 1]
+         must be 0.
+  nsyms: The number of symbols in the alphabet.
+         This should be at most 16.
+  Return: The decoded symbol s.*/
+static int od_ec_decode_cdf_q15(OdEcDec *dec, const uint16_t *icdf, int nsyms) {
+    OdEcWindow dif;
+    unsigned   r;
+    unsigned   c;
+    unsigned   u;
+    unsigned   v;
+    int        ret;
+    (void)nsyms;
+    dif         = dec->dif;
+    r           = dec->rng;
+    const int N = nsyms - 1;
+
+    assert(dif >> (OD_EC_WINDOW_SIZE - 16) < r);
+    assert(icdf[nsyms - 1] == OD_ICDF(CDF_PROB_TOP));
+    assert(32768U <= r);
+    assert(7 - EC_PROB_SHIFT - CDF_SHIFT >= 0);
+    c   = (unsigned)(dif >> (OD_EC_WINDOW_SIZE - 16));
+    v   = r;
+    ret = -1;
+    do {
+        u = v;
+        v = ((r >> 8) * (uint32_t)(icdf[++ret] >> EC_PROB_SHIFT) >>
+             (7 - EC_PROB_SHIFT - CDF_SHIFT));
+        v += EC_MIN_PROB * (N - ret);
+    } while (c < v);
+    assert(v < u);
+    assert(u <= r);
+    r = u - v;
+    dif -= (OdEcWindow)v << (OD_EC_WINDOW_SIZE - 16);
+    return od_ec_dec_normalize(dec, dif, r, ret);
+}
 
 /********************************************************************************************************************************/
 /********************************************************************************************************************************/
 /********************************************************************************************************************************/
 /********************************************************************************************************************************/
 // daalaboolreader.h from AOM
+
+/*This is meant to be a large, positive constant that can still be efficiently
+   loaded as an immediate (on platforms like ARM, for example).
+  Even relatively modest values like 100 would work fine.*/
+#define OD_EC_LOTS_OF_BITS (0x4000)
+
+/*The return value of od_ec_dec_tell does not change across an od_ec_dec_refill
+   call.*/
+static void od_ec_dec_refill(OdEcDec *dec) {
+    int                  s;
+    OdEcWindow           dif;
+    int16_t              cnt;
+    const unsigned char *bptr;
+    const unsigned char *end;
+    dif  = dec->dif;
+    cnt  = dec->cnt;
+    bptr = dec->bptr;
+    end  = dec->end;
+    s    = OD_EC_WINDOW_SIZE - 9 - (cnt + 15);
+    for (; s >= 0 && bptr < end; s -= 8, bptr++) {
+        /*Each time a byte is inserted into the window (dif), bptr advances and cnt
+       is incremented by 8, so the total number of consumed bits (the return
+       value of od_ec_dec_tell) does not change.*/
+        assert(s <= OD_EC_WINDOW_SIZE - 8);
+        dif ^= (OdEcWindow)bptr[0] << s;
+        cnt += 8;
+    }
+    if (bptr >= end) {
+        /*We've reached the end of the buffer. It is perfectly valid for us to need
+       to fill the window with additional bits past the end of the buffer (and
+       this happens in normal operation). These bits should all just be taken
+       as zero. But we cannot increment bptr past 'end' (this is undefined
+       behavior), so we start to increment dec->tell_offs. We also don't want
+       to keep testing bptr against 'end', so we set cnt to OD_EC_LOTS_OF_BITS
+       and adjust dec->tell_offs so that the total number of unconsumed bits in
+       the window (dec->cnt - dec->tell_offs) does not change. This effectively
+       puts lots of zero bits into the window, and means we won't try to refill
+       it from the buffer for a very long time (at which point we'll put lots
+       of zero bits into the window again).*/
+        dec->tell_offs += OD_EC_LOTS_OF_BITS - cnt;
+        cnt = OD_EC_LOTS_OF_BITS;
+    }
+    dec->dif  = dif;
+    dec->cnt  = cnt;
+    dec->bptr = bptr;
+}
+
+/*Initializes the decoder.
+  buf: The input buffer to use.
+  storage: The size in bytes of the input buffer.*/
+static void od_ec_dec_init(OdEcDec *dec, const unsigned char *buf, uint32_t storage) {
+    dec->buf       = buf;
+    dec->tell_offs = 10 - (OD_EC_WINDOW_SIZE - 8);
+    dec->end       = buf + storage;
+    dec->bptr      = buf;
+    dec->dif       = ((OdEcWindow)1 << (OD_EC_WINDOW_SIZE - 1)) - 1;
+    dec->rng       = 0x8000;
+    dec->cnt       = -15;
+    od_ec_dec_refill(dec);
+}
 
 typedef struct DaalaReader {
     const uint8_t *buffer;
@@ -142,11 +294,23 @@ typedef struct DaalaReader {
     uint8_t        allow_update_cdf;
 } DaalaReader_t;
 
-int aom_daala_reader_init(DaalaReader_t *r, const uint8_t *buffer, int size);
+static inline int svt_aom_daala_reader_init(DaalaReader_t *r, const uint8_t *buffer, int size) {
+    if (size && !buffer)
+        return 1;
+    r->buffer_end = buffer + size;
+    r->buffer     = buffer;
+    od_ec_dec_init(&r->ec, buffer, size);
+#if CONFIG_ACCOUNTING
+    r->accounting = NULL;
+#endif
+    return 0;
+}
 
-const uint8_t *aom_daala_reader_find_begin(DaalaReader_t *r);
+static inline const uint8_t *svt_aom_daala_reader_find_begin(DaalaReader_t *r) { return r->buffer; }
 
-const uint8_t *aom_daala_reader_find_end(DaalaReader_t *r);
+static inline const uint8_t *svt_aom_daala_reader_find_end(DaalaReader_t *r) {
+    return r->buffer_end;
+}
 
 static INLINE int aom_daala_read(DaalaReader_t *r, int prob) {
     int bit;
