+++ Source/Lib/Decoder/Codec/EbDecIntraPrediction.h
@@ -32,7 +32,7 @@ void svtav1_predict_intra_block(PartitionInfo *xd, int32_t plane, TxSize tx_size
                                 const PredictionMode mode, int32_t blk_mi_col_off,
                                 int32_t blk_mi_row_off, EbBitDepthEnum bit_depth, EbBool is_16bit);
 
-void cfl_store_tx(PartitionInfo *xd, CflCtx *cfl_ctx, int row, int col, TxSize tx_size,
+void svt_cfl_store_tx(PartitionInfo *xd, CflCtx *cfl_ctx, int row, int col, TxSize tx_size,
                   BlockSize bsize, EbColorConfig *cc, uint8_t *dst_buff, uint32_t dst_stride, EbBool is_16bit);
 
 #ifdef __cplusplus
