+++ Source/Lib/Decoder/Codec/EbDecParseInterBlock.h
@@ -46,7 +46,7 @@ extern void   av1_set_ref_frame(MvReferenceFrame *rf, int8_t ref_frame_type);
 
 void inter_block_mode_info(EbDecHandle *dec_handle, ParseCtxt *parse_ctxt, PartitionInfo *pi);
 
-void av1_find_mv_refs(EbDecHandle *dec_handle, PartitionInfo *pi, ParseCtxt *parse_ctx,
+void svt_av1_find_mv_refs(EbDecHandle *dec_handle, PartitionInfo *pi, ParseCtxt *parse_ctx,
                       MvReferenceFrame ref_frame, CandidateMv ref_mv_stack[][MAX_REF_MV_STACK_SIZE],
                       IntMv mv_ref_list[][MAX_MV_REF_CANDIDATES], IntMv global_mvs[2],
                       int16_t *mode_context, MvCount *mv_cnt);
